daemon: split up configuration API implementation into separate file
This reduces the size of `daemon.go`.

Signed-off by: Ian Vernon <ian@cilium.io>
