+++ daemon/daemon.go
@@ -27,10 +27,8 @@ import (
 	"time"
 
 	"github.com/cilium/cilium/api/v1/models"
-	. "github.com/cilium/cilium/api/v1/server/restapi/daemon"
 	health "github.com/cilium/cilium/cilium-health/launch"
 	"github.com/cilium/cilium/common"
-	"github.com/cilium/cilium/pkg/api"
 	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/cgroups"
 	"github.com/cilium/cilium/pkg/cidr"
@@ -89,7 +87,6 @@ import (
 	"github.com/cilium/cilium/pkg/trigger"
 	"github.com/cilium/cilium/pkg/workloads"
 
-	"github.com/go-openapi/runtime/middleware"
 	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sync/semaphore"
@@ -1449,120 +1446,6 @@ func changedOption(key string, value option.OptionSetting, data interface{}) {
 		proxy.ChangeLogLevel(logging.GetLevel(logging.DefaultLogger))
 	}
 	d.policy.BumpRevision() // force policy recalculation
-}
-
-type patchConfig struct {
-	daemon *Daemon
-}
-
-func NewPatchConfigHandler(d *Daemon) PatchConfigHandler {
-	return &patchConfig{daemon: d}
-}
-
-func (h *patchConfig) Handle(params PatchConfigParams) middleware.Responder {
-	log.WithField(logfields.Params, logfields.Repr(params)).Debug("PATCH /config request")
-
-	d := h.daemon
-
-	cfgSpec := params.Configuration
-
-	om, err := option.Config.Opts.Library.ValidateConfigurationMap(cfgSpec.Options)
-	if err != nil {
-		msg := fmt.Errorf("Invalid configuration option %s", err)
-		return api.Error(PatchConfigBadRequestCode, msg)
-	}
-
-	// Serialize configuration updates to the daemon.
-	option.Config.ConfigPatchMutex.Lock()
-	defer option.Config.ConfigPatchMutex.Unlock()
-
-	// Track changes to daemon's configuration
-	var changes int
-
-	// Only update if value provided for PolicyEnforcement.
-	if enforcement := cfgSpec.PolicyEnforcement; enforcement != "" {
-		switch enforcement {
-		case option.NeverEnforce, option.DefaultEnforcement, option.AlwaysEnforce:
-			// Update policy enforcement configuration if needed.
-			oldEnforcementValue := policy.GetPolicyEnabled()
-
-			// If the policy enforcement configuration has indeed changed, we have
-			// to regenerate endpoints and update daemon's configuration.
-			if enforcement != oldEnforcementValue {
-				log.Debug("configuration request to change PolicyEnforcement for daemon")
-				changes++
-				policy.SetPolicyEnabled(enforcement)
-			}
-
-		default:
-			msg := fmt.Errorf("Invalid option for PolicyEnforcement %s", enforcement)
-			log.Warn(msg)
-			return api.Error(PatchConfigFailureCode, msg)
-		}
-		log.Debug("finished configuring PolicyEnforcement for daemon")
-	}
-
-	changes += option.Config.Opts.ApplyValidated(om, changedOption, d)
-
-	log.WithField("count", changes).Debug("Applied changes to daemon's configuration")
-
-	if changes > 0 {
-		// Only recompile if configuration has changed.
-		log.Debug("daemon configuration has changed; recompiling base programs")
-		if err := d.compileBase(); err != nil {
-			msg := fmt.Errorf("Unable to recompile base programs: %s", err)
-			return api.Error(PatchConfigFailureCode, msg)
-		}
-		d.TriggerPolicyUpdates(true, "agent configuration update")
-	}
-
-	return NewPatchConfigOK()
-}
-
-type getConfig struct {
-	daemon *Daemon
-}
-
-func NewGetConfigHandler(d *Daemon) GetConfigHandler {
-	return &getConfig{daemon: d}
-}
-
-func (h *getConfig) Handle(params GetConfigParams) middleware.Responder {
-	log.WithField(logfields.Params, logfields.Repr(params)).Debug("GET /config request")
-
-	d := h.daemon
-
-	spec := &models.DaemonConfigurationSpec{
-		Options:           *option.Config.Opts.GetMutableModel(),
-		PolicyEnforcement: policy.GetPolicyEnabled(),
-	}
-
-	monitorState := d.monitorAgent.State()
-	status := &models.DaemonConfigurationStatus{
-		Addressing:       node.GetNodeAddressing(),
-		K8sConfiguration: k8s.GetKubeconfigPath(),
-		K8sEndpoint:      k8s.GetAPIServer(),
-		NodeMonitor:      &monitorState,
-		KvstoreConfiguration: &models.KVstoreConfiguration{
-			Type:    option.Config.KVStore,
-			Options: option.Config.KVStoreOpt,
-		},
-		Realized:     spec,
-		DeviceMTU:    int64(d.mtuConfig.GetDeviceMTU()),
-		RouteMTU:     int64(d.mtuConfig.GetRouteMTU()),
-		DatapathMode: models.DatapathMode(option.Config.DatapathMode),
-		IpvlanConfiguration: &models.IpvlanConfiguration{
-			MasterDeviceIndex: int64(option.Config.Ipvlan.MasterDeviceIndex),
-			OperationMode:     option.Config.Ipvlan.OperationMode,
-		},
-	}
-
-	cfg := &models.DaemonConfiguration{
-		Spec:   spec,
-		Status: status,
-	}
-
-	return NewGetConfigOK().WithPayload(cfg)
 }
 
 // listFilterIfs returns a map of interfaces based on the given filter.
